Setup data layer for getting the matches

+++ domain/src/main/java/ee/pw/edu/pl/domain/usecase/chat/people/ChatPerson.kt
@@ -1,13 +1,13 @@
 package ee.pw.edu.pl.domain.usecase.chat.people
 
 import android.os.Parcelable
-import kotlin.random.Random
 import kotlinx.parcelize.Parcelize
+import kotlin.random.Random
 
 @Parcelize
 data class ChatPerson(
 //    TODO remove default value when adding backend
-    val id: Long = Random.nextLong(),
+    val id: Int = Random.nextInt(),
     val name: String,
     val imageUrl: String,
 ) : Parcelable
